refactor(worklog-export): extract methods to new file
